requirements.txt
django
djangorestframework
djangorestframework_simplejwt
drf-yasg